+++ database/seeds/UserTableSeeder.php
@@ -1,32 +1,38 @@
 <?php
 
+use App\User;
+use App\UserPermission;
 use Illuminate\Database\Seeder;
+use Illuminate\Support\Facades\DB;
 
-class UserTableSeeder extends Seeder {
+class UserTableSeeder extends Seeder
+{
 
-	/**
-	 * Run the database seeds.
-	 *
-	 * @return void
-	 */
-	public function run()
-	{
-		DB::table('users')->delete();
+    /**
+     * Run the database seeds.
+     *
+     * @return void
+     */
+    public function run()
+    {
+        DB::table('users')->delete();
 
-		$thisIP = getHostByName(getHostName());
-		$testuser = User::create(array('username' => 'admin',
-							'email' => 'admin@admin.com',
-							'password' => Hash::make('admin'),
-							'created_ip' => $thisIP,
-							'last_ip' => $thisIP,
-							'created_by_user_id' => 1
-							));
+        $thisIP = getHostByName(getHostName());
+        $testuser = User::create([
+            'username' => 'admin',
+            'email' => 'admin@admin.com',
+            'password' => Hash::make('admin'),
+            'created_ip' => $thisIP,
+            'last_ip' => $thisIP,
+            'created_by_user_id' => 1
+        ]);
 
-		DB::table('user_permissions')->delete();
+        DB::table('user_permissions')->delete();
 
-		UserPermission::create(array('user_id' => $testuser->id,
-										'solder_full' => true
-										));
-	}
+        UserPermission::create([
+            'user_id' => $testuser->id,
+            'solder_full' => true
+        ]);
+    }
 
 }
