+++ tests/Unit/UrlUtilsTest.php
@@ -1,6 +1,9 @@
 <?php
 
+namespace Tests\Unit;
+
 use App\Libraries\UrlUtils;
+use Tests\TestCase;
 
 class UrlUtilsTest extends TestCase {
 
@@ -9,21 +12,21 @@ class UrlUtilsTest extends TestCase {
 		$json = UrlUtils::checkRemoteFile("https://httpbin.org/bytes/10", 5);
         $this->assertTrue(is_array($json));
 
-        $this->assertTrue(array_key_exists('success', $json));
+        $this->assertArrayHasKey('success', $json);
         $this->assertTrue($json['success']);
-        $this->assertTrue(array_key_exists('info', $json));
+        $this->assertArrayHasKey('info', $json);
         $this->assertEquals('200', $json['info']['http_code']);
 	}
 
     public function testCheckRemoteFileNon200()
 	{
-		$json = UrlUtils::checkRemoteFile("https://httpbin.org/status/404", 5);
+		$json = UrlUtils::checkRemoteFile("https://httpbin.org/status/404");
         $this->assertTrue(is_array($json));
 
-        $this->assertTrue(array_key_exists('success', $json));
+        $this->assertArrayHasKey('success', $json);
         $this->assertFalse($json['success']);
-        $this->assertTrue(array_key_exists('message', $json));
-        $this->assertTrue(array_key_exists('info', $json));
+        $this->assertArrayHasKey('message', $json);
+        $this->assertArrayHasKey('info', $json);
         $this->assertEquals('404', $json['info']['http_code']);
 	}
 
@@ -32,9 +35,9 @@ class UrlUtilsTest extends TestCase {
 		$json = UrlUtils::getHeaders("https://httpbin.org/bytes/10", 5);
         $this->assertTrue(is_array($json));
 
-        $this->assertTrue(array_key_exists('success', $json));
+        $this->assertArrayHasKey('success', $json);
         $this->assertTrue($json['success']);
-        $this->assertTrue(array_key_exists('info', $json));
+        $this->assertArrayHasKey('info', $json);
         $this->assertEquals('200', $json['info']['http_code']);
 	}
 
@@ -43,10 +46,10 @@ class UrlUtilsTest extends TestCase {
 		$json = UrlUtils::getHeaders("https://httpbin.org/status/404", 5);
         $this->assertTrue(is_array($json));
 
-        $this->assertTrue(array_key_exists('success', $json));
+        $this->assertArrayHasKey('success', $json);
         $this->assertFalse($json['success']);
-        $this->assertTrue(array_key_exists('message', $json));
-        $this->assertTrue(array_key_exists('info', $json));
+        $this->assertArrayHasKey('message', $json);
+        $this->assertArrayHasKey('info', $json);
         $this->assertEquals('404', $json['info']['http_code']);
 	}
 
@@ -55,11 +58,11 @@ class UrlUtilsTest extends TestCase {
 		$json = UrlUtils::get_remote_md5("https://httpbin.org/base64/dGVzdA==", 5);
         $this->assertTrue(is_array($json));
 
-        $this->assertTrue(array_key_exists('success', $json));
+        $this->assertArrayHasKey('success', $json);
         $this->assertTrue($json['success']);
-        $this->assertTrue(array_key_exists('md5', $json));
+        $this->assertArrayHasKey('md5', $json);
         $this->assertEquals('098f6bcd4621d373cade4e832627b4f6', $json['md5']);
-        $this->assertTrue(array_key_exists('filesize', $json));
+        $this->assertArrayHasKey('filesize', $json);
         $this->assertEquals('4', $json['filesize']);
 	}
 
@@ -68,10 +71,10 @@ class UrlUtilsTest extends TestCase {
 		$json = UrlUtils::get_remote_md5("https://httpbin.org/status/404", 5);
         $this->assertTrue(is_array($json));
 
-        $this->assertTrue(array_key_exists('success', $json));
+        $this->assertArrayHasKey('success', $json);
         $this->assertFalse($json['success']);
-        $this->assertTrue(array_key_exists('message', $json));
-        $this->assertTrue(array_key_exists('info', $json));
+        $this->assertArrayHasKey('message', $json);
+        $this->assertArrayHasKey('info', $json);
         $this->assertEquals('404', $json['info']['http_code']);
 	}
 }
